+++ Haromszogek/Haromszog.cs
@@ -5,9 +5,9 @@ namespace Haromszogek
 {
     internal class Haromszog
     {
-        private int aOldal;
-        private int bOldal;
-        private int cOldal;
+        private double aOldal;
+        private double bOldal;
+        private double cOldal;
         public double Terulet { get; private set; }
         public double Kerulet { get; private set; }
         public bool Szerkesztheto { get; private set; }
@@ -53,7 +53,7 @@ namespace Haromszogek
             return ker;
         }
 
-        public Haromszog(int aOldal, int bOldal, int cOldal)
+        public Haromszog(double aOldal, double bOldal, double cOldal)
         {
             this.aOldal = aOldal;
             this.bOldal = bOldal;
